Refactor: Layout 페이지 수정
