+++ src/commons/components/index.tsx
@@ -1,7 +1,6 @@
 import Card from "./Card/Card";
 import CardList from "./CardList/CardList";
-import Footer from "./Footer/Footer";
-import Nav from "./Nav/Nav";
+
 import SearchBar from "./SearchBar/SearchBar";
 
-export { Card, CardList, Footer, Nav, SearchBar };
+export { Card, CardList, SearchBar };
