+++ src/pages/Layout/components/Footer/Footer.module.css
@@ -1,4 +1,4 @@
-@import "../../../commons/styles/global.scss";
+@import "src/commons/styles/global.scss";
 
 footer {
   background: var(--the-julge-black, var(--color-black));
