+++ src/pages/Layout/components/Nav/Nav.module.scss
@@ -1,4 +1,4 @@
-@import "../../../commons/styles/global.scss";
+@import "src/commons/styles/global.scss";
 
 nav {
   width: 100%;
