+++ src/pages/Layout/components/Nav/Nav.tsx
@@ -2,7 +2,7 @@
 
 import { Link } from "react-router-dom";
 import styles from "./Nav.module.scss";
-import { ReactComponent as LogoSvg } from "../../../assets/images/logo.svg";
+import { ReactComponent as LogoSvg } from "src/assets/images/logo.svg";
 
 interface Props {
   profile: {
