+++ src/pages/Layout/components/index.tsx
@@ -1,0 +1,4 @@
+import Nav from "./Nav/Nav";
+import Footer from "./Footer/Footer";
+
+export { Nav, Footer };
